user_management_service/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
gunicorn==21.2.0
python-dotenv==1.0.1 